+++ app/models/base.js
