+++ lib/cache_management/base.js
@@ -5,8 +5,8 @@ const rootPrefix = '../..'
   , openStCacheKeys = openStCache.OpenSTCacheKeys
   , cacheImplementer = openStCache.cache
   , responseHelper = require(rootPrefix + '/lib/formatter/response')
-  , logger = require(rootPrefix+'/lib/logger/custom_console_logger')
-  , util = require(rootPrefix + '/lib/util')
+  , logger = require(rootPrefix+'/helpers/custom_console_logger')
+  , util = require(rootPrefix + '/lib/utils')
 ;
 
 /**
