+++ lib/cache_management/user_airdrop_detail.js
@@ -6,8 +6,8 @@ const rootPrefix = '../..'
   , userAirdropDetailModel = new UserAirdropDetailKlass()
   , coreConstants = require(rootPrefix + '/config/core_constants')
   , responseHelper = require(rootPrefix + '/lib/formatter/response')
-  , basicHelper = require(rootPrefix + '/helpers/basic')
-  , util = require(rootPrefix + '/lib/util')
+  , basicHelper = require(rootPrefix + '/helpers/basic_helper')
+  , util = require(rootPrefix + '/lib/utils')
 ;
 
 /**
