+++ lib/contract_interact/airdrop.js
@@ -140,11 +140,11 @@ Airdrop.prototype.pay = async function (
     options);
 
   const successCallback = function(receipt) {
-    return oThis.updateBalanceCacheOnReceipt(totalAmount, senderWorkerAddress, beneficiaryAddress, commissionBeneficiaryAddress, receipt);
+    return oThis.transactionHelper.updateBalanceCacheOnReceipt(oThis.brandedTokenAddress, totalAmount, senderWorkerAddress, beneficiaryAddress, commissionBeneficiaryAddress, receipt);
   };
 
   const failCallback = function(reason) {
-    return oThis.rollBackBalanceCache(senderWorkerAddress, totalAmount);
+    return oThis.transactionHelper.rollBackBalanceCache(oThis.brandedTokenAddress, senderWorkerAddress, totalAmount);
   };
 
   const params = {
@@ -161,7 +161,7 @@ Airdrop.prototype.pay = async function (
     errorCode: "l_ci_ad_p_3"
   };
 
-  return oThis.token.debitBalanceInCache(senderWorkerAddress, totalAmount).then(function(senderDebitCacheResponse) {
+  return oThis.transactionHelper.debitBalance(oThis.brandedTokenAddress, senderWorkerAddress, totalAmount).then(function(senderDebitCacheResponse) {
     if (senderDebitCacheResponse.isSuccess()) {
       return Promise.resolve(helper.performSend(params, returnType));
     } else {
@@ -169,7 +169,7 @@ Airdrop.prototype.pay = async function (
     }
   });
 
-},
+};
 
 /**
  * Get airdrop budget holder address of airdrop
