+++ lib/contract_interact/pricer.js
@@ -595,12 +595,12 @@ Pricer.prototype = {
       options);
 
     const successCallback = function(receipt) {
-      return oThis.updateBalanceCacheOnReceipt(totalAmount, senderAddress, beneficiaryAddress,
-          commissionBeneficiaryAddress, receipt);
+      return oThis.transactionHelper.updateBalanceCacheOnReceipt(oThis.brandedTokenAddress, totalAmount, senderAddress, beneficiaryAddress,
+        commissionBeneficiaryAddress, receipt);
     };
 
     const failCallback = function(reason) {
-      return oThis.rollBackBalanceCache(senderAddress, totalAmount);
+      return oThis.transactionHelper.rollBackBalanceCache(oThis.brandedTokenAddress, senderAddress, totalAmount);
     };
 
     const params = {
@@ -654,121 +654,6 @@ Pricer.prototype = {
     const totalEstimatedAmount = estimatedAmount.plus(estimatedCommisionAmount);
 
     return totalEstimatedAmount;
-  },
-
-  /**
-  * updateBalanceCacheOnReceipt
-  *
-  * @param {BigNumber} initialDebitAmount - amount that was debited
-  * @param {string} senderAddress - address of sender account
-  * @param {string} beneficiaryAddress - address of beneficiary account
-  * @param {string} commissionBeneficiaryAddress - address of commision beneficiary account
-  * @param {Object} transactionReceipt - transaction receipt
-  *
-  * @return {Promise}
-  *
-  */
-  updateBalanceCacheOnReceipt: function(
-    initialDebitAmount,
-    senderAddress,
-    beneficiaryAddress,
-    commissionBeneficiaryAddress,
-    transactionReceipt) {
-
-    const oThis = this;
-    // decode events
-    const decodedEvent = eventDecoder.perform(transactionReceipt, oThis.addressToNameMap);
-    var actualTransferAmount = new BigNumber(0);
-    var actualCommissionAmount = new BigNumber(0);
-
-    if (decodedEvent != undefined || decodedEvent != null) {
-      // get event data
-      const events =decodedEvent.formattedTransactionReceipt.eventsData;
-
-      if (events != undefined || events != null) {
-        // get whats the actual transfer amounts
-        for (var i = 0; i < events.length; i++) {
-          const eventData = events[i];
-
-          if (eventData.name === 'Payment') {
-            const paymentEvents = eventData.events;
-            for (var eventCount = 0; eventCount < paymentEvents.length; eventCount++) {
-              const paymentEventsData = paymentEvents[eventCount];
-              if (paymentEventsData.name === '_tokenAmount') {
-                actualTransferAmount = new BigNumber(paymentEventsData.value);
-              } else if (paymentEventsData.name === '_commissionTokenAmount') {
-                actualCommissionAmount = new BigNumber(paymentEventsData.value);
-              }
-            }
-          }
-
-        }
-      }
-    }
-
-
-    // update cache for beneficiary (credit)
-    const creditBeneficiaryBalancePromise = new Promise(function (onResolve, onReject) {
-      if (actualTransferAmount > 0) {
-        oThis.transactionHelper.creditBalance(oThis.brandedTokenAddress, beneficiaryAddress, actualTransferAmount).then(function(beneficiaryCreditCacheResponse) {              
-          onResolve(beneficiaryCreditCacheResponse);
-        });
-      } else {
-        onResolve(responseHelper.successWithData({}));
-      }
-    });
-
-    // update cache for commission beneficiary (credit)
-    const creditCommissionBeneficiaryBalancePromise = new Promise(function (onResolve, onReject) {
-      if (actualCommissionAmount > 0) {
-        oThis.transactionHelper.creditBalance(oThis.brandedTokenAddress, commissionBeneficiaryAddress, actualCommissionAmount).then(function(commissionBeneficiaryCreditCacheResponse) {              
-          Promise.resolve(commissionBeneficiaryCreditCacheResponse);
-        });
-      } else {
-        onResolve(responseHelper.successWithData({}));
-      }
-    });
-
-    // adjustment cache for spender
-    const actualTotalAmount = actualTransferAmount.plus(actualCommissionAmount);
-
-    const adjustmentBalancePromise = new Promise(function (onResolve, onReject) {
-      if (actualTotalAmount.gt(initialDebitAmount)) {
-        const adjustedAmount = actualTotalAmount.minus(initialDebitAmount);
-        oThis.transactionHelper.debitBalance(oThis.brandedTokenAddress, senderAddress, adjustedAmount).then(function(senderAdjustedCreditCacheResponse) {
-          onResolve(senderAdjustedCreditCacheResponse);
-        });
-      } else if (actualTotalAmount.lt(initialDebitAmount)) {
-        const adjustedAmount = initialDebitAmount.minus(actualTotalAmount);        
-        oThis.transactionHelper.creditBalance(oThis.brandedTokenAddress, senderAddress, adjustedAmount).then(function(senderAdjustedDebitCacheResponse) {
-          onResolve(senderAdjustedDebitCacheResponse);
-        });
-      } else {
-        onResolve(responseHelper.successWithData({}));
-      }
-    });
-
-    return Promise.all([creditBeneficiaryBalancePromise, creditCommissionBeneficiaryBalancePromise, adjustmentBalancePromise]);
-
-  },
-
-  /**
-  * rollBackBalanceCache
-  *
-  * @param {string} senderAddress - address of sender account
-  * @param {BigNumber} totalAmount - amount that needs to be credited
-  *
-  * @return {Promise}
-  *
-  */
-  rollBackBalanceCache: function(senderAddress, totalAmount) {
-
-    const oThis = this;
-    return new Promise(function (onResolve, onReject) {
-      oThis.transactionHelper.creditBalance(senderAddress, totalAmount).then(function(senderCreditCacheResponse) {
-        onResolve(senderCreditCacheResponse);
-      });
-    });
   },
