+++ lib/transaction_helper.js
@@ -7,13 +7,14 @@
  * @module lib/cache_management/transaction_cache_helper
  *
  */
-const rootPrefix = '../..'
+const rootPrefix = '..'
   , responseHelper = require(rootPrefix + '/lib/formatter/response')
-  , BalanceCache =  require(rootPrefix + '/lib/cache_management/balance_cache')
-  , Token = require('./branded_token')
+  , BalanceCache =  require(rootPrefix + '/lib/cache_management/balance')
+  , Token = require(rootPrefix + '/lib/contract_interact/branded_token')
   , BigNumber = require('bignumber.js')
   , UserAirdropDetailCacheKlass = require(rootPrefix + '/lib/cache_management/user_airdrop_detail')
-  , UserAirdropDetailModelKlass = require(rootPrefix + '/app/model/user_airdrop_detail')
+  , UserAirdropDetailModelKlass = require(rootPrefix + '/app/models/user_airdrop_detail')
+  , eventDecoder = require(rootPrefix + '/lib/web3/events/decoder')
   ;
 
 
@@ -194,5 +195,124 @@ TransactionHelper.prototype = {
       owner,
       bigAmount
     );
+  },
+
+
+  /**
+  * updateBalanceCacheOnReceipt
+  *
+  * @param {string} brandedTokenAddress - branded token address
+  * @param {BigNumber} initialDebitAmount - amount that was debited
+  * @param {string} senderAddress - address of sender account
+  * @param {string} beneficiaryAddress - address of beneficiary account
+  * @param {string} commissionBeneficiaryAddress - address of commision beneficiary account
+  * @param {Object} transactionReceipt - transaction receipt
+  *
+  * @return {Promise}
+  *
+  */
+  updateBalanceCacheOnReceipt: function(
+    brandedTokenAddress,
+    initialDebitAmount,
+    senderAddress,
+    beneficiaryAddress,
+    commissionBeneficiaryAddress,
+    transactionReceipt) {
+
+    const oThis = this;
+    // decode events
+    const decodedEvent = eventDecoder.perform(transactionReceipt, oThis.addressToNameMap);
+    var actualTransferAmount = new BigNumber(0);
+    var actualCommissionAmount = new BigNumber(0);
+
+    if (decodedEvent != undefined || decodedEvent != null) {
+      // get event data
+      const events =decodedEvent.formattedTransactionReceipt.eventsData;
+
+      if (events != undefined || events != null) {
+        // get whats the actual transfer amounts
+        for (var i = 0; i < events.length; i++) {
+          const eventData = events[i];
+
+          if (eventData.name === 'Payment') {
+            const paymentEvents = eventData.events;
+            for (var eventCount = 0; eventCount < paymentEvents.length; eventCount++) {
+              const paymentEventsData = paymentEvents[eventCount];
+              if (paymentEventsData.name === '_tokenAmount') {
+                actualTransferAmount = new BigNumber(paymentEventsData.value);
+              } else if (paymentEventsData.name === '_commissionTokenAmount') {
+                actualCommissionAmount = new BigNumber(paymentEventsData.value);
+              }
+            }
+          }
+
+        }
+      }
+    }
+
+
+    // update cache for beneficiary (credit)
+    const creditBeneficiaryBalancePromise = new Promise(function (onResolve, onReject) {
+      if (actualTransferAmount > 0) {
+        oThis.creditBalance(brandedTokenAddress, beneficiaryAddress, actualTransferAmount).then(function(beneficiaryCreditCacheResponse) {              
+          onResolve(beneficiaryCreditCacheResponse);
+        });
+      } else {
+        onResolve(responseHelper.successWithData({}));
+      }
+    });
+
+    // update cache for commission beneficiary (credit)
+    const creditCommissionBeneficiaryBalancePromise = new Promise(function (onResolve, onReject) {
+      if (actualCommissionAmount > 0) {
+        oThis.creditBalance(brandedTokenAddress, commissionBeneficiaryAddress, actualCommissionAmount).then(function(commissionBeneficiaryCreditCacheResponse) {              
+          Promise.resolve(commissionBeneficiaryCreditCacheResponse);
+        });
+      } else {
+        onResolve(responseHelper.successWithData({}));
+      }
+    });
+
+    // adjustment cache for spender
+    const actualTotalAmount = actualTransferAmount.plus(actualCommissionAmount);
+
+    const adjustmentBalancePromise = new Promise(function (onResolve, onReject) {
+      if (actualTotalAmount.gt(initialDebitAmount)) {
+        const adjustedAmount = actualTotalAmount.minus(initialDebitAmount);
+        oThis.debitBalance(brandedTokenAddress, senderAddress, adjustedAmount).then(function(senderAdjustedCreditCacheResponse) {
+          onResolve(senderAdjustedCreditCacheResponse);
+        });
+      } else if (actualTotalAmount.lt(initialDebitAmount)) {
+        const adjustedAmount = initialDebitAmount.minus(actualTotalAmount);
+        oThis.creditBalance(brandedTokenAddress, senderAddress, adjustedAmount).then(function(senderAdjustedDebitCacheResponse) {
+          onResolve(senderAdjustedDebitCacheResponse);
+        });
+      } else {
+        onResolve(responseHelper.successWithData({}));
+      }
+    });
+
+    return Promise.all([creditBeneficiaryBalancePromise, creditCommissionBeneficiaryBalancePromise, adjustmentBalancePromise]);
+
+  },
+
+  /**
+  * rollBackBalanceCache
+  *
+  * @param {string} brandedTokenAddress - branded token address
+  * @param {string} senderAddress - address of sender account
+  * @param {BigNumber} totalAmount - amount that needs to be credited
+  *
+  * @return {Promise}
+  *
+  */
+  rollBackBalanceCache: function(brandedTokenAddress, senderAddress, totalAmount) {
+
+    const oThis = this;
+    return new Promise(function (onResolve, onReject) {
+      oThis.creditBalance(brandedTokenAddress, senderAddress, totalAmount).then(function(senderCreditCacheResponse) {
+        onResolve(senderCreditCacheResponse);
+      });
+    });
   }
 };
